reorganise and add some tests for authorization rules
